+++ Robin/src/Robin/ImGui/imgui_layer.cpp
@@ -82,7 +82,7 @@ namespace Robin
 	{
 		ImGuiIO& io = ImGui::GetIO();
 		application& app = application::get();
-		io.DisplaySize = ImVec2(app.get_window().get_width(), app.get_window().get_height());
+		io.DisplaySize = ImVec2((float)app.get_window().get_width(), (float)app.get_window().get_height());
 
 		// rendering 
 		ImGui::Render();
